Tidy up CI scripts and remove those no longer in use
