--- ci/install_houdini2.sh
@@ -1,9 +1,0 @@
-#!/usr/bin/env bash
-
-set -e
-
-if [ -d "hou" ]; then
-    # move hou tarball into top-level and untar
-    cp hou/hou.tar.gz .
-    tar -xzf hou.tar.gz
-fi
