--- ci/install_jemalloc.sh
@@ -1,5 +1,0 @@
-#!/usr/bin/env bash
-
-set -ex
-
-apt-get install -y libjemalloc-dev
